Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -11,7 +11,7 @@ class CreateCostumes < ActiveRecord::Migration[5.1]
   def change
     create_table :costumes do |t|
       t.string :name,
-      t.float :price,
+      t.decimal :price,
       t.string :image_url
     end
   end
